Refactored the library and improved folder structure
