+++ composer.json
@@ -2,20 +2,24 @@
     "name": "fullpipe/twig-webpack-extension",
     "description": "Inject your webpack entry points into twig templates with easy.",
     "type": "library",
-    "keywords": ["webpack", "twig", "extension"],
-    "require": {
-        "twig/twig": "^1.20 || ^2.0"
-    },
     "license": "MIT",
+    "keywords": [
+        "webpack",
+        "twig",
+        "extension"
+    ],
     "authors": [
         {
             "name": "Eugene Bravov",
             "email": "eugene.bravov@gmail.com"
         }
     ],
+    "require": {
+        "twig/twig": "^1.20 || ^2.0"
+    },
     "autoload": {
         "psr-4": {
-            "Fullpipe\\Twig\\Extension\\Webpack\\": ""
+            "Fullpipe\\TwigWebpackExtension\\": "src/Fullpipe/TwigWebpackExtension/"
         }
     }
 }
