+++ src/Fullpipe/TwigWebpackExtension/WebpackExtension.php
@@ -1,34 +1,16 @@
 <?php
 
-namespace Fullpipe\Twig\Extension\Webpack;
+namespace Fullpipe\TwigWebpackExtension;
 
-/**
- * WebpackExtension.
- */
+use Fullpipe\TwigWebpackExtension\TokenParser\EntryTokenParserCss;
+use Fullpipe\TwigWebpackExtension\TokenParser\EntryTokenParserJs;
+
 class WebpackExtension extends \Twig_Extension
 {
-    /**
-     * @var string
-     */
     protected $manifestFile;
-
-    /**
-     * @var string
-     */
     protected $publicPathJs;
-
-    /**
-     * @var string
-     */
     protected $publicPathCss;
 
-    /**
-     * Constructor.
-     *
-     * @param string $manifestFile  absolute path to your manifest.json
-     * @param string $publicPathJs  your webpack output.publicPath
-     * @param string $publicPathCss your webpack output.publicPath
-     */
     public function __construct($manifestFile, $publicPathJs = '/js/', $publicPathCss = '/css/')
     {
         $this->manifestFile = $manifestFile;
@@ -36,17 +18,11 @@ class WebpackExtension extends \Twig_Extension
         $this->publicPathCss = $publicPathCss;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getName()
     {
         return 'fullpipe.extension.webpack';
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getTokenParsers()
     {
         return [
